requirements-test.txt
pytest
mock
vcrpy
pytest-cov
coveralls
